extract interface for CUBRIDOID
